refactor: remove _Computed alias and unify signal import paths
- Remove Computed as _Computed alias in app/styles.py
- Unify SignalBase, Computed, and computed_property imports to public webcompy.signal path
- Mark tasks 5-6 complete
