+++ packages/webcompy/src/webcompy/elements/typealias/_element_property.py
@@ -3,7 +3,7 @@ from typing import Any, TypeAlias
 
 from webcompy.elements._dom_objs import DOMEvent
 from webcompy.elements.types._abstract import ElementAbstract
-from webcompy.signal._base import SignalBase
+from webcompy.signal import SignalBase
 
 ElementChildren: TypeAlias = ElementAbstract | SignalBase[Any] | str | None
 AttrValue: TypeAlias = SignalBase[Any] | str | int | bool
